+++ src/event/listener/KeyDownEventListener.test.ts
@@ -1,11 +1,11 @@
 import { beforeEach, describe, expect, it, vitest } from 'vitest'
 import EventHandler from '../handler/EventHandler'
 import { fireEvent, getByRole, waitFor } from '@testing-library/dom'
-import PointerUpEventListener from './PointerUpEventListener'
 import { nop } from '../../utils/nop'
 import createElementInJSDOM from '../../test-utils/createElementInJSDOM'
+import KeyDownEventListener from './KeyDownEventListener'
 
-describe('ClickEventListener', () => {
+describe('KeyDownEventListener', () => {
   describe('listener', () => {
     const eventHandler = new EventHandler('aaa-111', 0, nop)
     const runSpy = vitest.spyOn(eventHandler, 'run')
@@ -14,19 +14,19 @@ describe('ClickEventListener', () => {
       vitest.restoreAllMocks()
     })
 
-    it('calls listener when click event been fired by mouse', async () => {
+    it('calls listener when key down event been fired', async () => {
       const { element, domWindow } = createElementInJSDOM(
         `
                 <div>
-                    <button class="size-lg"/>
+                    <input type="checkbox" id="checkbox1" name="checkbox1"/>
                 </div>`,
         'div',
       )
 
-      new PointerUpEventListener(eventHandler, domWindow).init()
-      const button = await waitFor(() => getByRole(element, 'button'))
+      new KeyDownEventListener(eventHandler, domWindow).init()
+      const button = await waitFor(() => getByRole(element, 'checkbox'))
 
-      fireEvent.pointerUp(button)
+      fireEvent.keyDown(button)
 
       await waitFor(() => {
         expect(runSpy).toHaveBeenCalledOnce()