+++ src/event/listener/KeyUpEventListener.test.ts
@@ -32,26 +32,5 @@ describe('KeyUpEventListener', () => {
         expect(runSpy).toHaveBeenCalledOnce()
       })
     })
-
-    it('does not call listener if up key is not space', async () => {
-      const { element, domWindow } = createElementInJSDOM(
-        `
-                <div>
-                    <input type="checkbox" id="checkbox1" name="checkbox1"/>
-                </div>`,
-        'div',
-      )
-
-      new KeyUpEventListener(eventHandler, domWindow).init()
-      const button = await waitFor(() => getByRole(element, 'checkbox'))
-
-      fireEvent.keyUp(button, { code: 'enter' })
-
-      await waitFor(() => {}, { timeout: 500 })
-
-      await waitFor(() => {
-        expect(runSpy).toBeCalledTimes(0)
-      })
-    })
   })
 })